chore(types): move types to shared directory
