+++ apps/web/src/app/features/chatbots/components/UserMessage.tsx
@@ -1,4 +1,4 @@
-import type { Message } from '../../types';
+import type { Message } from '../types';
 
 interface UserMessageProps {
     message: Message;
